refactor(#347): move CompletionBadge to components/common
The component serves both training (set/exercise/session kinds) and
nutrition (meal/day kinds) surfaces since #329 added the nutrition
variants. The components/training/ path was misleading. components/common/
is the project's cross-domain bucket — same neighbour as DayNoteInput.

No behaviour change. Pure file move + import path update across the 5
consumer files (SetRow, ExerciseCardHeader, MealBlock, TrainingPlanPage,
NutritionPlanPage).
