+++ web/src/components/common/CompletionBadge.tsx
